requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
scapy==2.7.0
xgboost==3.2.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
sqlalchemy==2.0.25
httpx==0.26.0
python-dotenv==1.0.0
psutil==7.2.2
pydantic==2.10.0
pydantic-core==2.27.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
python-multipart==0.0.9
